Implement unkeyed decoding for NSControl, NSResponder, NSTextField and NSView (darlinghq/darling#724)

+++ AppKit/nib.subproj/NSIBObjectData.m
@@ -319,8 +319,8 @@
 					[key release];
 				}
 
-				unsigned long long nextOid;
-				[coder decodeValueOfObjCType: @encode(unsigned long long)
+				int nextOid;
+				[coder decodeValueOfObjCType: @encode(int)
 					at: &nextOid];
 				_nextOid = nextOid;
